Fold island-originated actions into one wrapper
Three parallel cases (islandEntryTapped, islandHandOffTapped,
islandRunWorkflowTapped) plus a context-menu router all expressed the
same rule: collapse the roster, surface Prowl, then run the sidebar
action. Replace them with `island(Action)`. The reducer forwards the
wrapped action and collapses only when it presents Prowl UI
(`surfacesProwl`); AppFeature surfaces the window in one case instead
of four, and Open Prowl now calls the lifecycle client synchronously
like its siblings. The island navigation tests move out of the quit
test file into their own.
